+++ packages/client-node/__tests__/integration/node_summary.test.ts
@@ -30,48 +30,69 @@ whenOnEnv(
       values: jsonValues,
       format: 'JSONEachRow',
     })
-    expect(insertSummary).toEqual({
-      read_rows: '5',
-      read_bytes: jasmine.any(String),
-      written_rows: '5',
-      written_bytes: jasmine.any(String),
-      total_rows_to_read: '0',
-      result_rows: '5',
-      result_bytes: jasmine.any(String),
-      elapsed_ns: jasmine.any(String),
-    })
+    expect(insertSummary).toEqual(
+      jasmine.objectContaining({
+        read_rows: '5',
+        read_bytes: jasmine.any(String),
+        written_rows: '5',
+        written_bytes: jasmine.any(String),
+        total_rows_to_read: '0',
+        result_rows: '5',
+        result_bytes: jasmine.any(String),
+        elapsed_ns: jasmine.any(String),
+      }),
+    )
+    assertRealTimeMicroseconds(insertSummary)
 
     const { summary: execSummary } = await client.exec({
-      query: `INSERT INTO ${tableName} SELECT * FROM ${tableName}`,
+      query: `INSERT INTO ${tableName}
+              SELECT *
+              FROM ${tableName}`,
     })
-    expect(execSummary).toEqual({
-      read_rows: '5',
-      read_bytes: jasmine.any(String),
-      written_rows: '5',
-      written_bytes: jasmine.any(String),
-      total_rows_to_read: '5',
-      result_rows: '5',
-      result_bytes: jasmine.any(String),
-      elapsed_ns: jasmine.any(String),
-    })
+    expect(execSummary).toEqual(
+      jasmine.objectContaining({
+        read_rows: '5',
+        read_bytes: jasmine.any(String),
+        written_rows: '5',
+        written_bytes: jasmine.any(String),
+        total_rows_to_read: '5',
+        result_rows: '5',
+        result_bytes: jasmine.any(String),
+        elapsed_ns: jasmine.any(String),
+      }),
+    )
+    assertRealTimeMicroseconds(execSummary)
   })
 
   it('should provide summary for command', async () => {
     const { summary } = await client.command({
-      query: `INSERT INTO ${tableName} VALUES (144, 'Hello', [2, 4]), (255, 'World', [3, 5])`,
+      query: `INSERT INTO ${tableName}
+              VALUES (144, 'Hello', [2, 4]),
+                     (255, 'World', [3, 5])`,
       clickhouse_settings: {
         wait_end_of_query: 1,
       },
     })
-    expect(summary).toEqual({
-      read_rows: '2',
-      read_bytes: jasmine.any(String),
-      written_rows: '2',
-      written_bytes: jasmine.any(String),
-      total_rows_to_read: '0',
-      result_rows: '2',
-      result_bytes: jasmine.any(String),
-      elapsed_ns: jasmine.any(String),
-    })
+    expect(summary).toEqual(
+      jasmine.objectContaining({
+        read_rows: '2',
+        read_bytes: jasmine.any(String),
+        written_rows: '2',
+        written_bytes: jasmine.any(String),
+        total_rows_to_read: '0',
+        result_rows: '2',
+        result_bytes: jasmine.any(String),
+        elapsed_ns: jasmine.any(String),
+      }),
+    )
+    assertRealTimeMicroseconds(summary)
   })
+
+  function assertRealTimeMicroseconds(summary: any) {
+    // FIXME: remove this condition after 24.9 is released
+    if (process.env['CLICKHOUSE_VERSION'] === 'head') {
+      expect(summary.real_time_microseconds).toBeDefined()
+      expect(summary.real_time_microseconds).toMatch(/^\d+$/)
+    }
+  }
 })